fix latest commit
I've been doing a lot more "Merge" requests lately, which gets bundled
in with the commits in the activity payload, but we're only looking at
the latest commit, which will be the merge one. Instead flatten all
commit messages, but keep the same filters.

Update some basic info in README

Update checkout to latest